requirements.txt
asyncio~=3.4.3
starlette~=0.46.2
langgraph~=0.4.5
langchain-core~=0.3.60
pydantic~=2.11.4
langchain-openai~=0.3.17
PyYAML~=6.0.2
pytest~=8.3.5
pytest-asyncio~=0.26.0
uvicorn~=0.34.2
PyJWT~=2.10.1
httpx~=0.28.1
jwcrypto~=1.5.6
httpx-sse~=0.4.0
sse-starlette~=2.3.5
typing_extensions~=4.13.2